Commit and refactor
